Replace task macro with trait
While the macro provided good ergonomics, it hardcoded configuration parameters and hid a lot of how a task is actually created. It was also possible to create tasks from anywhere in the dependency tree without control of the PXROS application developer.
With the task trait, all configurations are user accessible and task creation is done in one single point.