Replace task macro with trait (#3)
* Replace task macro with trait

While the macro provided good ergonomics, it hardcoded configuration parameters and hid a lot of how a task is actually created. It was also possible to create tasks from anywhere in the dependency tree without control of the PXROS application developer.
With the task trait, all configurations are user accessible and task creation is done in one single point.

* Update CI and xtask check

* Update examples and docs

* Switch to builder for `TaskCreationConfig`

A builder allows more ergonomic overriding of single parts of the configuration without providing many combinations of override methods. The current implementation prevents accidentally overriding the same configuration twice, thus making it more user-friendly and misuse-resistant.

---------

Co-authored-by: vE5li <ve5li@tuta.io>